Move predicate test util classes from contentpal to contentpal-testing. #99
