+++ contentpal-testing/src/main/java/org/dmfs/android/contentpal/testing/predicates/ValueArgument.java
@@ -14,29 +14,44 @@
  * limitations under the License.
  */
 
-package org.dmfs.android.contentpal.predicates.utils;
+package org.dmfs.android.contentpal.testing.predicates;
+
+import android.support.annotation.NonNull;
 
 import org.dmfs.android.contentpal.Predicate;
-import org.dmfs.iterables.decorators.DelegatingIterable;
-import org.dmfs.iterables.decorators.Mapped;
-import org.dmfs.iterators.Function;
+import org.dmfs.optional.Absent;
 import org.dmfs.optional.Optional;
 
 
 /**
+ * A simple constant argument with no back reference.
+ *
  * @author Marten Gajda
  */
-public final class BackReferences extends DelegatingIterable<Optional<Integer>>
+public final class ValueArgument implements Predicate.Argument
 {
-    public BackReferences(Iterable<Predicate.Argument> arguments)
+    private final Object mValue;
+
+
+    public ValueArgument(@NonNull Object value)
     {
-        super(new Mapped<>(arguments, new Function<Predicate.Argument, Optional<Integer>>()
-        {
-            @Override
-            public Optional<Integer> apply(Predicate.Argument argument)
-            {
-                return argument.backReference();
-            }
-        }));
+
+        mValue = value;
+    }
+
+
+    @NonNull
+    @Override
+    public String value()
+    {
+        return mValue.toString();
+    }
+
+
+    @NonNull
+    @Override
+    public Optional<Integer> backReference()
+    {
+        return Absent.absent();
     }
 }
