+++ contentpal-testing/src/main/java/org/dmfs/android/contentpal/testing/predicates/Values.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2017 dmfs GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dmfs.android.contentpal.testing.predicates;
+
+import org.dmfs.android.contentpal.Predicate;
+import org.dmfs.iterables.decorators.DelegatingIterable;
+import org.dmfs.iterables.decorators.Mapped;
+import org.dmfs.iterators.Function;
+
+
+/**
+ * @author Marten Gajda
+ */
+public final class Values extends DelegatingIterable<String>
+{
+    public Values(Iterable<Predicate.Argument> arguments)
+    {
+        super(new Mapped<>(arguments, new Function<Predicate.Argument, String>()
+        {
+            @Override
+            public String apply(Predicate.Argument argument)
+            {
+                return argument.value();
+            }
+        }));
+    }
+}
